re-organize windows named pipe code
